Parser keeps track of local variable dictionaries
The parser now keeps track of the set of local variables in each function, as
well as which of those variables are used non-locally (with the 'nonlocal'
keyword in a nested scope).  These are stored in the AST for functions and
generators.

darcs-hash:20091103204826-8b57f-2f88a1d60df98e85e541ee7a8623ed74c6a30ff5.gz