Unresolved type variables default to Any
Ignore-this: 775ef2c01d98952674d412a3347d19b5

It's possible for a type to be completely unconstrained.  This happens for
variables that are both undefined and unused on some control flow paths.
Since there are no constraints, type inference creates a fresh type variable
for them, but the variable doesn't get a value from anywhere.  The result is
an undefined variable error in System F.

This patch fixes that by keeping track of all type variables that are
mentioned, filtering out all type variables that are defined, and defaulting
the remaining type variables to the distinguished type 'Any', which is
isomorphic to the unit type.

darcs-hash:20100219024234-8b57f-972ccff379c74a37cc56f58f77b94e0be0841f41.gz