Allow casting to a larger effect type even if the effect type escapes
Ignore-this: b1a14eff9ee50f1164b1a41cd5bfa5a9

When a function with a value parameter is cast to a function with a region
parameter, we know that the effect will not escape even if the type says it
does.  Therefore we can eliminate one constraint from the subtype test.
This is necessary to handle stream polymorphism, because the effect appears
as part of the stream's side effect.

This problem manifested in stream code, where a function with type
(int -> stream(EmpE, int)) was used at the type
(a in r -> <{r}> stream({r}, a)).  It looks like the parameter region 'r'
escapes, which is not allowed.  However, the wrapper function we generate
has the type (int in r -> <{r}> stream(EmpE, int)), which is safe (note that
the stream's effect is EmpE, not {r}).

darcs-hash:20100804161455-8b57f-f82a0844d633470c717aae8485f93ece2bebd402.gz