Change meaning of varIsExternal
Ignore-this: 6a85814cb40977a74ba64ed97cbdbc0a

The field of Var used to be True if the variable was defined outside the
current module.  However, builtin variables didn't have the right value since,
when they're created, it's not known whether they're defined in the current
module.

The new meaning is True if the variable is visible outside the current module.
Check the module's definitions to know whether a variable is defined in the
current module.

darcs-hash:20101031001025-11156-af07a90ba29278e60bd67d9cecd01affbbdbe5f1.gz