Completed the change over to shape-indexed streams
Ignore-this: e62751603b1100136af9e625f16f3eab

The types of map and zip_N have changed, and the rewriter is changed
to work with shape-indexed streams.

In this patch, the generation of stream types is slightly different,
leading to a more verbose syntax.  Unfortunately, we need that to be
able to write all kinds of kernels.

darcs-hash:20110218211731-11156-785f0b6bb486135dcc810090f5dbce305f1af3b3.gz