Eliminate local objects that are deconstructed by case statements
Ignore-this: e4e0ede6e9deb15c7874d20de74a3163

This patch enables floating on fields of local variables when the variable is a
reference variable.  This is basically the same transformation that we do for
value and boxed values, but we enable it for reference objects
only if we're guaranteed to benefit.  (Performing the transformation
indiscriminately would be very costly, because it introduces deep-copying of
objects.)

darcs-hash:20110303215526-11156-698459b19048e5348062b69b7544162a12ab267b.gz