Compute abstract values of functions
Ignore-this: 4f0a01a668c31865ca14d1979d7ff870

Functions are now interpreted to abstract values during the simplifier.
The purpose of interpretation is to put initializer functions on the same
footing as ordinary values.

The new code is not exercised by our tests.  Expect bugs in the code.
Since this is a low-priority change and it's passing the tests, I'm checking
it in.

darcs-hash:20110715173011-11156-c4304e3f8e69ac10e79dd8e740673dba8e80f3c6.gz