Major changes to builtin library and bugfixes
Ignore-this: 4750b8572a63790649326337c76ae9a2

The built-in library was changed to implement the new container
data types.  The changes exposed some new bugs, some of which are
fixed in this patch.

The functionality in this patch is incomplete; some functionality
is missing and some changes may be only partly implemented.

darcs-hash:20110823171246-11156-d5fdb3fe9bfd5c9f565194733deff371baadc307.gz