Build fewer closures in closure conversion
Ignore-this: 208a4221dd8d963022084b2563764bbe
If a function has only direct calls, then its closure is not used.
This patch detects when a function is the target of direct calls only.
In such cases, it does not allow the function to be closure-captured and
does not generate closures for the function.

darcs-hash:20120719000757-11156-6ee2012d603f958fbdcd53c6e769a8e11faa43bb.gz