Adjusted the test for dependent variables in constraints
Ignore-this: 3163e722e40a2b9bc40c45d57eb12b1d
This adjustment is necessary to work with the new defaulting rules for
container types.  Constraints of the form 'shape a ~ dim2' should not
lead to 'a' being dependent on other variables, so that 'a' can be a
candidate for defaulting.

darcs-hash:20121006025354-11156-4900fd4dc0cb9bc50995d3245e9672d9d9fa4a82.gz