Backend uses cursors
Ignore-this: 2feb6c1445398afd277ef0e6766f0048
In preparation for implementing marshaling, bare object references
are now implemented with cursors.
A cursor is a pair of pointers: one to the head of an object, and one
to some interior section that is of interest.

darcs-hash:20130427223626-11156-700c4edfe9330669f7156fb7f597b2ea0180bc8b.gz