Use if-of-equality-test to substitute variables
Ignore-this: 5a193cb49032b09430bf6c2797b04d8d
In an expression of the form

> let b = (x == y) in ... if b then f x y else ...

in the "then" part, we know that x == y, so we
substitute x for y.  This is helpful because it can trigger
further simplification of arithmetic that uses both x and y.

darcs-hash:20130903060348-11156-c264a30bd77d861751527cc4a873ee11a257b8de.gz